Be conservative at say 24 Mpps for a single core
We have benchmarks showing between 24-26 Mpps for a single core.

In the XDP DROP test we had to enable rx_cqe_compress=on, to avoid
a PCIe BW limitation, and this setting comes with a slight overhead
of decompressing the descriptor, thus lowering the per core performance.
Thus, lets report on this lower number in the article.

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>